+++ src/RegularTest.java
@@ -1,4 +1,3 @@
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
 
 public class RegularTest extends SudokuTest{
@@ -20,7 +19,7 @@ public class RegularTest extends SudokuTest{
                 {2, 0, 5, 4, 0, 1, 6, 0, 3}
         };
 
-        int[][] solution =
+        long[][] solution =
                 {
                         {9, 1, 4, 2, 6, 7, 8, 3, 5},
                         {5, 8, 6, 1, 3, 9, 4, 7, 2},
@@ -36,7 +35,7 @@ public class RegularTest extends SudokuTest{
         AbstractPuzzle sudoku = new RegularSudoku.RegularSudokuBuilder()
                 .withGivens(givens)
                 .build();
-        int[][] solverSolution = sudoku.solve();
+        long[][] solverSolution = sudoku.solve();
 
         compareSolutions(solverSolution, solution);
     }
@@ -55,7 +54,7 @@ public class RegularTest extends SudokuTest{
                 {0, 0, 0, 1, 0, 0, 5, 0, 0}
         };
 
-        int[][] solution =
+        long[][] solution =
                 {
                         {5, 9, 8, 2, 7, 1, 3, 4, 6},
                         {7, 4, 2, 6, 5, 3, 8, 9, 1},
@@ -71,7 +70,7 @@ public class RegularTest extends SudokuTest{
         AbstractPuzzle sudoku = new RegularSudoku.RegularSudokuBuilder()
                 .withGivens(givens)
                 .build();
-        int[][] solverSolution = sudoku.solve();
+        long[][] solverSolution = sudoku.solve();
 
         compareSolutions(solverSolution, solution);
     }
@@ -87,7 +86,7 @@ public class RegularTest extends SudokuTest{
                 {0, 0, 0, 0, 0, 0}
         };
 
-        int[][] solution = {
+        long[][] solution = {
                 {6, 2, 3, 4, 5, 1},
                 {5, 4, 1, 3, 2, 6},
                 {4, 3, 2, 1, 6, 5},
